pyproject.toml
[project]
name = "basic-agent"
version = "0.1.0"
description = "A reusable Python library for building LLM-powered agents"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.77.0",
    "jinja2>=3.1.6",
    "openai>=2.16.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "psycopg[binary]>=3.2.13",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
